Add files via upload
This is a data visualization  of Oil prices since January 1st, 1990. The following code was used to iron out issues, including that the "." in the data that denotes the change between dollars and cents was not being read properly by the compiler. This full code will not be useful in all circumstances, but will be important when the compiler can not read the price data accurately.